+++ XIVAuras/changelog.md
@@ -1,6 +1,9 @@
 # Version 0.2.0.1
 - New fonts are now automatically added to the font list on import
 - Added 'Solid Color' Icon type
+- Added Ability to dynamically track 'Adjusted Actions'
+    - An adjusted action is an ability that changes/upgrades during combat.
+    - Examples: Standard Step -> Standard Finish, Gallows -> Cross Reaping, etc.
 - Fixed labels appearing on top of other Icons when stacked
 
 # Version 0.2.0.0